Refactor the design of the code: UserController can only accessed the db
