+++ code/app/src/main/java/com/cmput301f21t26/habittracker/ui/habit/TodayHabitFragment.java
@@ -1,6 +1,7 @@
 package com.cmput301f21t26.habittracker.ui.habit;
 
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -53,6 +54,7 @@ public class TodayHabitFragment extends Fragment {
         navController = Navigation.findNavController(getActivity(), R.id.nav_host_fragment_activity_main);
 
         todayHabitList = (ArrayList<Habit>) UserController.getCurrentUser().getTodayHabits();
+        Log.d("Number", String.valueOf(todayHabitList.size()));
 
         rvListener = new HabitAdapter.RecyclerViewClickListener() {
             @Override
